chore: move LIP.lua to src/tools/
